requirements.txt
azure-functions
azure-data-tables
bs4
datetime
logging
requests

typing
lxml
fastapi
pydantic